pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transcriptor"
version = "0.1.0"
description = "Voice-to-text transcription for Linux using Whisper"
requires-python = ">=3.10"
dependencies = [
    "faster-whisper>=1.0.0",
    "sounddevice>=0.4.6",
    "pynput>=1.7.6",
    "pystray>=0.19.0",
    "Pillow>=10.0.0",
    "PyGObject>=3.42.0",
    "numpy>=1.24.0",
]

[project.scripts]
transcriptor = "transcriptor.main:main"

[tool.setuptools.packages.find]
include = ["transcriptor*"]
